Fixed Edit Page redirect to Profile Page after saving changes. (#20)
Bug 18 - After 'Save Changes' in profile, it stays in the same page. Typically its better to redirect to the Landing page.

Before:
Edit Page did not redirect to Profile Page once we updated the details and clicked the 'Save Changes' button.

After:
Now, it successfully redirects to the Profile Page after clicking the 'Save Changes' button.