1. Change the format of time display in 'Add Event.' Taking 9 am as an example: Before, it would display as "9:0". After the update, it will display as "09:00"."
2. When we add a event to firebase through the app, set 'interest' field in the event to 0 instead of 1

3. change notifyDataSetChanged() to notifyItemChanged(position) to improve the performance.

4. Fix the bug:
Before: When the bottom sheet is expanded, clicking the 'heart button' on the event list does not reflect the result on the 'heart button' on the bottom sheet for the same selected event. Additionally, the area displaying the profile image and the number of users interested in the event does not display the expected results when clicking the 'heart button' on either the event list or the bottom sheet.

After: The 'heart button' behaves consistently across the event list and bottom sheet. The area displaying the profile image and the number of users interested in the event accurately reflects the results accordingly.